Revert "Merge remote-tracking branch 'origin/develop' into develop"
This reverts commit e79d2dff3e2adef6d3fc6407772144d53a51d06f, reversing
changes made to 51c8b589141d0a427a65f9ce74da287fb3410dd9.

+++ packages/contracts-bedrock/contracts/test/Encoding.t.sol
@@ -91,7 +91,7 @@ contract Encoding_Test is CommonTest {
         uint64 _gas,
         bool isCreate,
         bytes memory _data,
-        uint64 _logIndex
+        uint256 _logIndex
     ) external {
         Types.UserDepositTransaction memory t = Types.UserDepositTransaction(
             _from,
